backend/requirements.txt
# --- CORE FRAMEWORK ---
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
python-multipart==0.0.9
python-dotenv==1.0.1

# --- AI & INTELLIGENCE ---
google-genai==0.3.0
tenacity==8.2.3

# --- DATA SCIENCE & MATH ---
pandas==2.2.0
numpy==1.26.4

# --- WEB SCRAPING ---
httpx==0.26.0
beautifulsoup4==4.12.3

# --- TESTING ---
pytest==8.0.0